requirements.txt
EasyProcess==0.2.3
Pillow==5.2.0
pkg-resources==0.0.0
PyVirtualDisplay==0.2.1
selenium==3.14.0
urllib3==1.23
